AI Progress
Can make AI paths in the editor (still a work in progress) and give them to AI Agents for them to follow. You Can create a path by dragging the Prefab "EnemyPatrolPath" to the scene and using the inspector buttons to add.remove points (edit the points in the scene). When finished give the path to the enemy Agent.